resume-screener-nlp/requirements.txt
streamlit
scikit-learn
pdfminer.six
